+++ packages/kobber-components/src/index.web-components.ts
@@ -16,3 +16,4 @@ export { Heading } from "./text/heading/Heading";
 export { Carousel } from "./carousel/Carousel";
 export { HorizontalLayout } from "./layouts/horizontal-layout/HorizontalLayout";
 export { HorizontalLayoutColumn } from "./layouts/horizontal-layout/HorizontalLayoutColumn";
+export { RadioGroup, RadioInput, RadioInputControl } from "./radio";