+++ packages/kobber-components/src/radio/radioGroup/RadioGroup.ts
@@ -12,9 +12,9 @@ import {
   valueMissingValidityState,
 } from "../../base/internal/form";
 import componentStyles from "../../base/styles/component.styles";
-import type RadioInput from "../radioInput/RadioInput";
+import type { RadioInput } from "../radioInput/RadioInput";
 import { radioGroupName, radioInputName, radioGroupHorizontalClassName, GroupProps } from "../Radio.core";
-import { buttonGroupStyles } from "./RadioGroup.styles";
+import { radioGroupStyles } from "./RadioGroup.styles";
 
 /**
  * @summary Radio groups are used to group multiple [radios](/components/radio) or [radio buttons](/components/radio-button) so they function as a single form control.
@@ -36,13 +36,13 @@ import { buttonGroupStyles } from "./RadioGroup.styles";
 type Props = GroupProps & ShoelaceFormControl;
 
 @customElement(radioGroupName)
-export default class RadioGroup extends ShoelaceElement implements Props {
+export class RadioGroup extends ShoelaceElement implements Props {
   disabled?: boolean | undefined;
   defaultChecked?: boolean | undefined;
   input!: HTMLInputElement;
   validationMessage: string = "";
 
-  static styles: CSSResultGroup = [componentStyles, buttonGroupStyles];
+  static styles: CSSResultGroup = [componentStyles, radioGroupStyles];
 
   protected readonly formControlController = new FormControlController(this);
   private readonly hasSlotController = new HasSlotController(this, "help-text", "label");