+++ packages/kobber-audio-recorder/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gyldendal/kobber-audio-recorder
 
+## 0.3.77
+
+### Patch Changes
+
+- Correct kobber-radio components
+
 ## 0.3.76
 
 ### Patch Changes